requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
Django==3.0.6
django-autoslug==1.9.7
django-bootstrap4==1.1.1
Pillow==7.1.2
pytz==2020.1
soupsieve==2.0
sqlparse==0.3.1
